+++ Content.Server/Speech/EntitySystems/BackwardsAccentSystem.cs
@@ -1,4 +1,5 @@
 using Content.Server.Speech.Components;
+using Content.Shared.Speech;
 
 namespace Content.Server.Speech.EntitySystems
 {
