+++ Content.Server/Speech/EntitySystems/DamagedSiliconAccentSystem.cs
@@ -4,6 +4,7 @@ using Content.Server.PowerCell;
 using Content.Shared.Speech.Components;
 using Content.Shared.Damage;
 using Content.Shared.FixedPoint;
+using Content.Shared.Speech;
 using Robust.Shared.Random;
 
 namespace Content.Server.Speech.EntitySystems;
