+++ Content.Server/Speech/EntitySystems/GermanAccentSystem.cs
@@ -1,7 +1,8 @@
 using System.Text;
-using Content.Server.Speech.Components;
-using Robust.Shared.Random;
 using System.Text.RegularExpressions;
+using Content.Server.Speech.Components;
+using Content.Shared.Speech;
+using Robust.Shared.Random;
 
 namespace Content.Server.Speech.EntitySystems;
