+++ Content.Server/Speech/EntitySystems/LizardAccentSystem.cs
@@ -1,5 +1,6 @@
 ﻿using System.Text.RegularExpressions;
 using Content.Server.Speech.Components;
+using Content.Shared.Speech;
 
 namespace Content.Server.Speech.EntitySystems;
