+++ Content.Server/Speech/EntitySystems/MumbleAccentSystem.cs
@@ -1,6 +1,7 @@
 using Content.Server.Chat.Systems;
 using Content.Server.Speech.Components;
 using Content.Shared.Chat.Prototypes;
+using Content.Shared.Speech;
 using Content.Shared.Speech.Components;
 
 namespace Content.Server.Speech.EntitySystems;
