+++ Content.Server/Speech/EntitySystems/PirateAccentSystem.cs
@@ -1,5 +1,6 @@
 using System.Linq;
 using Content.Server.Speech.Components;
+using Content.Shared.Speech;
 using Robust.Shared.Random;
 using System.Text.RegularExpressions;
