+++ Content.Server/Speech/EntitySystems/SkeletonAccentSystem.cs
@@ -1,5 +1,6 @@
 ﻿using System.Text.RegularExpressions;
 using Content.Server.Speech.Components;
+using Content.Shared.Speech;
 using Robust.Shared.Random;
 
 namespace Content.Server.Speech.EntitySystems;
