+++ Content.Server/Speech/EntitySystems/SpanishAccentSystem.cs
@@ -1,5 +1,6 @@
 using System.Text;
 using Content.Server.Speech.Components;
+using Content.Shared.Speech;
 
 namespace Content.Server.Speech.EntitySystems
 {
