+++ Content.Shared/Speech/AccentEvent.cs
@@ -1,0 +1,21 @@
+﻿namespace Content.Shared.Speech;
+
+public sealed class AccentGetEvent : EntityEventArgs
+{
+    /// <summary>
+    ///     The entity to apply the accent to.
+    /// </summary>
+    public EntityUid Entity { get; }
+
+    /// <summary>
+    ///     The message to apply the accent transformation to.
+    ///     Modify this to apply the accent.
+    /// </summary>
+    public string Message { get; set; }
+
+    public AccentGetEvent(EntityUid entity, string message)
+    {
+        Entity = entity;
+        Message = message;
+    }
+}
